pyproject.toml
[project]
name = "TeamUno_Python"
version = "0.2.0"
readme = "README.md"
requires-python = ">=3.13"

[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
asgiref = "^3.8.1"
Django = "^5.1.6"
django-cors-headers = "^4.7.0"
djangorestframework = "^3.15.2"
load_dotenv = "^0.1.0"
Pillow = "^11.1.0"
psycopg2 = "^2.9.10"
python = "^3.13"
sqlparse = "^0.5.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
